# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Hoteles",
        "description": "API para gestionar hoteles, habitaciones, huéspedes y servicios",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/hotel": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Obtener lista de hoteles",
                "description": "Devuelve una lista de hoteles con la posibilidad de filtrar por nombre, dirección, teléfono, email y website, y paginar los resultados.",
                "operationId": "42e38e1ec9a77688ed78f16868c19a31",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre del hotel",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "adress",
                        "in": "query",
                        "description": "Filtrar por dirección del hotel",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phoneNumber",
                        "in": "query",
                        "description": "Filtrar por número de teléfono del hotel",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filtrar por email del hotel",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "website",
                        "in": "query",
                        "description": "Filtrar por sitio web del hotel",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de hoteles obtenida"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Crear un nuevo hotel",
                "description": "Crea un nuevo hotel con los datos introducidos",
                "operationId": "e4cd0a9fd83fe69e1de0b7278c89a329",
                "requestBody": {
                    "description": "Datos necesarios para crear un hotel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "adress": {
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "website": {
                                        "type": "string"
                                    },
                                    "services": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel creado"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/hotel/{id}": {
            "get": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Obtener un hotel",
                "operationId": "834e722d4a653eb33c419e11980be20e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel encontrado"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Actualizar un hotel",
                "description": "Actualiza un hotel con los datos introducidos",
                "operationId": "aaa283c9b6f6d4bf9baed2a64f9705f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos que se pueden actualizar del hotel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "adress": {
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "website": {
                                        "type": "string"
                                    },
                                    "services": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel actualizado"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hoteles"
                ],
                "summary": "Eliminar un hotel",
                "description": "Elimina un hotel por su ID.",
                "operationId": "c028370bd4871a5de33d5e2e2e8af789",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel eliminado"
                    },
                    "404": {
                        "description": "Hotel no encontrado"
                    }
                }
            }
        },
...
    },
...
}